requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.2
pandas==3.0.0
pillow==12.1.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
